feat: Enhance image processing with AI integration and buffer uploads to Cloudinary

+++ BackEnd/src/utils/Cloudinary.ts
@@ -1,8 +1,9 @@
 import { v2 as cloudinary } from "cloudinary";
 import dotenv from "dotenv";
 import * as fs from "fs";
+import { upload } from "../middleware/multer.middleware";
+import { ApiError } from "./ApiError";
 dotenv.config();
-
 
 cloudinary.config({
   cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
@@ -10,7 +11,7 @@ cloudinary.config({
   api_secret: process.env.CLOUDINARY_API_SECRET,
 });
 
-export const UploadFile = async (imagePath:string) => {
+export const UploadFile = async (imagePath: string) => {
   // Use the uploaded file's name as the asset's public ID and
   // allow overwriting the asset with new versions
   const options = {
@@ -23,15 +24,20 @@ export const UploadFile = async (imagePath:string) => {
     // Upload the image
     const result = await cloudinary.uploader.upload(imagePath, options);
     console.log("File is succeccfully uploaded on cloudinary server ", result);
+    if(result && fs.existsSync(imagePath)){
+      fs.unlinkSync(imagePath);
+    }
     return result;
   } catch (error) {
     console.error("src :: utils :: Cloudinary", error);
-    fs.unlinkSync(imagePath);
+    if(fs.existsSync(imagePath)){
+      fs.unlinkSync(imagePath);
+    }
     throw error;
   }
 };
 
-export const DeleteFile = async (publicId:string) => {
+export const DeleteFile = async (publicId: string) => {
   try {
     const result = await cloudinary.uploader.destroy(publicId);
     return result;
@@ -40,3 +46,25 @@ export const DeleteFile = async (publicId:string) => {
     throw error;
   }
 };
+
+export async function uploadBufferToCloudinary(buffer: Buffer) {
+  try {
+    console.log("Uploading buffer to Cloudinary...",buffer);
+    const uploadResult = await new Promise((resolve, reject) => {
+      const stream = cloudinary.uploader.upload_stream(
+        { folder:"ai-generated" },
+        (error, result) => {
+          if (error) reject(error);
+          else resolve(result);
+        }
+      );
+
+      stream.end(buffer);
+    });
+
+    return uploadResult as any;
+  } catch (err: any) {
+    console.error("Cloudinary upload error:", err);
+    throw new ApiError(500, "Failed to upload image: " + err.message);
+  }
+}
